src/requirements.txt
pyyaml
pypresence
requests
beautifulsoup4
